moves custom exceptions to different package
