+++ src/main/java/de/rabea/exceptions/FileWritingException.java
@@ -1,4 +1,4 @@
-package de.rabea.request;
+package de.rabea.exceptions;
 
 public class FileWritingException extends RuntimeException {
