+++ src/test/java/de/rabea/exceptions/ShaComputingExceptionTest.java
@@ -1,8 +1,8 @@
-package de.rabea.request;
+package de.rabea.exceptions;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 public class ShaComputingExceptionTest {
