+++ src/test/java/de/rabea/request/Sha1EncoderTest.java
@@ -2,7 +2,7 @@ package de.rabea.request;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 public class Sha1EncoderTest {
